+++ MoltenVK/MoltenVK/GPUObjects/MVKSwapchain.h
@@ -28,8 +28,6 @@
 
 class MVKWatermark;
 
-@class MVKBlockObserver;
-
 
 #pragma mark -
 #pragma mark MVKSwapchain
@@ -76,19 +74,8 @@ public:
 	/** Releases swapchain images. */
 	VkResult releaseImages(const VkReleaseSwapchainImagesInfoEXT* pReleaseInfo);
 
-	/** Returns whether the parent surface is now lost and this swapchain must be recreated. */
-	bool getIsSurfaceLost() { return _surfaceLost; }
-
-	/** Returns whether this swapchain is optimally sized for the surface. */
-	bool hasOptimalSurface();
-
 	/** Returns the status of the surface. Surface loss takes precedence over sub-optimal errors. */
-	VkResult getSurfaceStatus() {
-		if (_device->getConfigurationResult() != VK_SUCCESS) { return _device->getConfigurationResult(); }
-		if (getIsSurfaceLost()) { return VK_ERROR_SURFACE_LOST_KHR; }
-		if ( !hasOptimalSurface() ) { return VK_SUBOPTIMAL_KHR; }
-		return VK_SUCCESS;
-	}
+	VkResult getSurfaceStatus();
 
 	/** Adds HDR metadata to this swapchain. */
 	void setHDRMetadataEXT(const VkHdrMetadataEXT& metadata);
@@ -118,31 +105,28 @@ protected:
 						  VkSwapchainPresentScalingCreateInfoEXT* pScalingInfo,
 						  uint32_t imgCnt);
 	void initSurfaceImages(const VkSwapchainCreateInfoKHR* pCreateInfo, uint32_t imgCnt);
-	void releaseLayer();
-	void releaseUndisplayedSurfaces();
+	bool getIsSurfaceLost();
+	bool hasOptimalSurface();
 	uint64_t getNextAcquisitionID();
-    void willPresentSurface(id<MTLTexture> mtlTexture, id<MTLCommandBuffer> mtlCmdBuff);
     void renderWatermark(id<MTLTexture> mtlTexture, id<MTLCommandBuffer> mtlCmdBuff);
     void markFrameInterval();
-	void recordPresentTime(const MVKImagePresentInfo& presentInfo, uint64_t actualPresentTime = 0);
+	void beginPresentation(const MVKImagePresentInfo& presentInfo);
+	void endPresentation(const MVKImagePresentInfo& presentInfo, uint64_t actualPresentTime = 0);
 
-	CAMetalLayer* _mtlLayer = nil;
+	MVKSurface* _surface = nullptr;
     MVKWatermark* _licenseWatermark = nullptr;
 	MVKSmallVector<MVKPresentableSwapchainImage*, kMVKMaxSwapchainImageCount> _presentableImages;
 	MVKSmallVector<VkPresentModeKHR, 2> _compatiblePresentModes;
 	static const int kMaxPresentationHistory = 60;
 	VkPastPresentationTimingGOOGLE _presentTimingHistory[kMaxPresentationHistory];
 	std::atomic<uint64_t> _currentAcquisitionID = 0;
-    MVKBlockObserver* _layerObserver = nil;
 	std::mutex _presentHistoryLock;
-	std::mutex _layerLock;
 	uint64_t _lastFrameTime = 0;
 	VkExtent2D _mtlLayerDrawableExtent = {0, 0};
 	uint32_t _currentPerfLogFrameCount = 0;
 	uint32_t _presentHistoryCount = 0;
 	uint32_t _presentHistoryIndex = 0;
 	uint32_t _presentHistoryHeadIndex = 0;
-	std::atomic<bool> _surfaceLost = false;
 	bool _isDeliberatelyScaled = false;
 };